+++ lib/ohai/application.rb
@@ -15,12 +15,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+require 'chef-config/workstation_config_loader'
 require 'ohai'
 require 'ohai/log'
 require 'mixlib/cli'
 
 class Ohai::Application
   include Mixlib::CLI
+
+  option :config_file,
+    :short       => "-c CONFIG",
+    :long        => "--config CONFIG",
+    :description => "A configuration file to use",
+    :proc => lambda { |path| File.expand_path(path, Dir.pwd) }
 
   option :directory,
     :short       => "-d DIRECTORY",
@@ -74,6 +81,7 @@ class Ohai::Application
     @attributes = parse_options
     @attributes = nil if @attributes.empty?
 
+    load_workstation_config
     Ohai::Config.merge_deprecated_config
     Ohai.config.merge!(config)
     if Ohai.config[:directory]
@@ -110,6 +118,18 @@ class Ohai::Application
     def exit!(msg, err = -1)
       Ohai::Log.debug(msg)
       Process.exit err
+    end
+  end
+
+  private
+  def load_workstation_config
+    config_loader = ChefConfig::WorkstationConfigLoader.new(
+      config[:config_file], Ohai::Log
+    )
+    begin
+      config_loader.load
+    rescue ChefConfig::ConfigurationError => config_error
+      Ohai::Application.fatal!(config_error.message)
     end
   end
 end